Move class list item
