+++ typescript/web/src/components/class-selection-popover/class-list-item/__stories__/class-list-item.tsx
@@ -1,7 +1,7 @@
 import { omit } from "lodash/fp";
 
 import { ClassListItem } from "../class-list-item";
-import { chakraDecorator } from "../../../utils/chakra-decorator";
+import { chakraDecorator } from "../../../../utils/chakra-decorator";
 
 export default {
   title: "web/Class list item",
