+++ typescript/web/src/components/class-selection-popover/class-list-item/__tests__/class-list-item.tsx
