+++ typescript/web/src/components/class-selection-popover/class-list-item/class-list-item.tsx
